backend/requirements.txt
fastapi
uvicorn
typing_extensions
langchain
langchain-community
openai
python-dotenv
requests 